Refactor the pay elements form to be reusable
We want to use it for editing productive time as well.
